feat: Provide extension point to declare a DAP server
See #838
