+++ src/main/java/com/redhat/devtools/lsp4ij/dap/DAPDebugRunner.java
@@ -30,6 +30,7 @@ import com.intellij.xdebugger.XDebugSession;
 import com.intellij.xdebugger.XDebuggerManager;
 import com.redhat.devtools.lsp4ij.dap.configurations.DAPCommandLineState;
 import com.redhat.devtools.lsp4ij.dap.configurations.DAPRunConfiguration;
+import com.redhat.devtools.lsp4ij.dap.configurations.DAPRunConfigurationBase;
 import com.redhat.devtools.lsp4ij.installation.CommandLineUpdater;
 import com.redhat.devtools.lsp4ij.installation.ConsoleProvider;
 import com.redhat.devtools.lsp4ij.installation.ServerInstallationStatus;
@@ -62,7 +63,7 @@ public class DAPDebugRunner extends AsyncProgramRunner<RunnerSettings> {
      */
     @Override
     public boolean canRun(@NotNull String executorId, @NotNull RunProfile profile) {
-        return profile instanceof DAPRunConfiguration dap && dap.canRun(executorId);
+        return profile instanceof DAPRunConfigurationBase<?> dap && dap.canRun(executorId);
     }
 
     /**
