+++ src/main/java/com/redhat/devtools/lsp4ij/dap/breakpoints/DAPBreakpointType.java
@@ -24,9 +24,9 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 /**
- *  Debug Adapter Protocol (DAP) breakpoint type.
+ * Default Debug Adapter Protocol (DAP) breakpoint type.
  */
-public class DAPBreakpointType extends XLineBreakpointType<DAPBreakpointProperties> {
+public class DAPBreakpointType extends DAPBreakpointTypeBase<DAPBreakpointProperties> {
 
     private static final String BREAKPOINT_ID = "dap-breakpoint";
 
@@ -40,32 +40,4 @@ public class DAPBreakpointType extends XLineBreakpointType<DAPBreakpointProperti
         return new DAPBreakpointProperties();
     }
 
-    @Override
-    public boolean canPutAt(@NotNull VirtualFile file,
-                            int line,
-                            @NotNull Project project) {
-        return DebugAdapterManager.getInstance().isDebuggableFile(file, project);
-    }
-
-    @Override
-    public @Nullable XDebuggerEditorsProvider getEditorsProvider(@NotNull XLineBreakpoint<DAPBreakpointProperties> breakpoint,
-                                                                 @NotNull Project project) {
-        // Returns a non-null XDebuggerEditorsProvider to support Condition in the breakpoint type.
-        FileType fileType = null;
-        var file = LSPIJUtils.findResourceFor(breakpoint.getFileUrl());
-        if (file != null) {
-            if (!canSupportConditionBreakpoint(file, project)) {
-                return null;
-            }
-            fileType = file.getFileType();
-        }
-        return new DAPDebuggerEditorsProvider(fileType, null);
-    }
-
-    private boolean canSupportConditionBreakpoint(VirtualFile file, @NotNull Project project) {
-        // TODO: manage a settings to know if the given file which is associated to a DAP server
-        // can support condition breakpoint point
-        // At this step, the DAP server cannot be started, so we need to manage a DAP server settings for that.
-        return true;
-    }
 }
