+++ src/main/java/com/redhat/devtools/lsp4ij/dap/descriptors/DebugAdapterDescriptor.java
@@ -23,10 +23,19 @@ import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.Key;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.util.net.NetUtils;
+import com.intellij.xdebugger.XDebugSession;
+import com.intellij.xdebugger.breakpoints.XBreakpointProperties;
+import com.intellij.xdebugger.breakpoints.XLineBreakpoint;
 import com.redhat.devtools.lsp4ij.dap.DAPDebugProcess;
 import com.redhat.devtools.lsp4ij.dap.DebugMode;
+import com.redhat.devtools.lsp4ij.dap.breakpoints.DAPBreakpointHandler;
+import com.redhat.devtools.lsp4ij.dap.breakpoints.DAPBreakpointHandlerBase;
+import com.redhat.devtools.lsp4ij.dap.breakpoints.DAPBreakpointProperties;
 import com.redhat.devtools.lsp4ij.dap.client.DAPClient;
+import com.redhat.devtools.lsp4ij.dap.client.LaunchUtils;
+import com.redhat.devtools.lsp4ij.dap.configurations.DAPRunConfigurationOptions;
 import com.redhat.devtools.lsp4ij.dap.configurations.DebuggableFile;
+import com.redhat.devtools.lsp4ij.dap.configurations.options.AttachConfigurable;
 import com.redhat.devtools.lsp4ij.dap.definitions.DebugAdapterServerDefinition;
 import com.redhat.devtools.lsp4ij.internal.IntelliJPlatformUtils;
 import com.redhat.devtools.lsp4ij.internal.StringUtils;
@@ -223,7 +232,24 @@ public abstract class DebugAdapterDescriptor implements DebuggableFile {
      * @return the strategy to use to know when DAP server is started and DAP client can connect to it.
      */
     @NotNull
-    public abstract ServerReadyConfig getServerReadyConfig(@NotNull DebugMode debugMode);
+    public ServerReadyConfig getServerReadyConfig(@NotNull DebugMode debugMode) {
+        if (options instanceof AttachConfigurable dapOptions) {
+            if (debugMode == DebugMode.ATTACH) {
+                String address = LaunchUtils.resolveAttachAddress(dapOptions.getAttachAddress(), getDapParameters());
+                int port = LaunchUtils.resolveAttachPort(dapOptions.getAttachPort(), getDapParameters());
+                return new ServerReadyConfig(address, port);
+            }
+        }
+        if (options instanceof DAPRunConfigurationOptions dapOptions) {
+            var strategy = dapOptions.getDebugServerWaitStrategy();
+            return switch (strategy) {
+                case TIMEOUT -> new ServerReadyConfig(dapOptions.getConnectTimeout());
+                case TRACE -> new ServerReadyConfig(dapOptions.getNetworkAddressExtractor());
+                default -> new ServerReadyConfig(0);
+            };
+        }
+        return new ServerReadyConfig(500);
+    }
 
     public abstract @Nullable FileType getFileType();
 
@@ -244,5 +270,9 @@ public abstract class DebugAdapterDescriptor implements DebuggableFile {
     @Override
     public boolean isDebuggableFile(@NotNull VirtualFile file, @NotNull Project project) {
         return serverDefinition != null && serverDefinition.isDebuggableFile(file, project);
+    }
+
+    public @NotNull DAPBreakpointHandlerBase<?> createBreakpointHandler(@NotNull XDebugSession session, Project project) {
+        return new DAPBreakpointHandler(session, this, project);
     }
 }
